List decoder reads list of integers and strings.

+++ src/main/java/com/ffbit/bencode/IntegerDecoder.java
@@ -102,7 +102,7 @@ public class IntegerDecoder implements Decoder<Integer> {
         sb.setLength(0);
     }
 
-    public boolean isApplicable(byte b) {
+    public boolean isApplicable(int b) {
         return b == PREF;
     }
